feat(api): modernize router layer [CODX-API-REF-312]
