+++ tests/api/test_error_contract.py
